reorganize module hierarchy
