+++ pytissueoptics/rayscattering/__init__.py
@@ -1,15 +1,13 @@
 from .photon import Photon
 from .materials import ScatteringMaterial
+from .energyLogging import EnergyLogger
 from .source import PencilPointSource, IsotropicPointSource, DirectionalSource, DivergentSource
-from .statistics import Stats
-from .energyLogger import EnergyLogger
-from .viewer import Viewer, PointCloudStyle, Visibility
-from .views import ViewGroup, Direction
-from .views.defaultViews import View2DProjection, View2DProjectionX, View2DProjectionY, View2DProjectionZ, \
+from .display.viewer import Viewer, PointCloudStyle, Visibility, ViewGroup, Direction
+from .display.views import View2DProjection, View2DProjectionX, View2DProjectionY, View2DProjectionZ, \
     View2DSurface, View2DSurfaceX, View2DSurfaceY, View2DSurfaceZ, View2DSlice, View2DSliceX, View2DSliceY, View2DSliceZ
 
 __all__ = ["Photon", "ScatteringMaterial", "PencilPointSource", "IsotropicPointSource", "DirectionalSource",
-              "DivergentSource", "Stats", "EnergyLogger", "Viewer", "PointCloudStyle", "Visibility", "ViewGroup",
+              "DivergentSource", "EnergyLogger", "Viewer", "PointCloudStyle", "Visibility", "ViewGroup",
                 "Direction", "View2DProjection", "View2DProjectionX", "View2DProjectionY", "View2DProjectionZ",
                 "View2DSurface", "View2DSurfaceX", "View2DSurfaceY", "View2DSurfaceZ", "View2DSlice", "View2DSliceX",
                 "View2DSliceY", "View2DSliceZ"]
