+++ pytissueoptics/rayscattering/display/utils/__init__.py
@@ -1,0 +1,2 @@
+from .direction import Direction, DEFAULT_X_VIEW_DIRECTIONS, DEFAULT_Y_VIEW_DIRECTIONS, DEFAULT_Z_VIEW_DIRECTIONS
+from .volumeSlicer import VolumeSlicer
