+++ pytissueoptics/rayscattering/display/views/__init__.py
@@ -1,0 +1,4 @@
+from .view2D import ViewGroup, View2D
+from .viewFactory import ViewFactory
+from .defaultViews import View2DProjection, View2DProjectionX, View2DProjectionY, View2DProjectionZ, \
+    View2DSurface, View2DSurfaceX, View2DSurfaceY, View2DSurfaceZ, View2DSlice, View2DSliceX, View2DSliceY, View2DSliceZ
