+++ pytissueoptics/rayscattering/energyLogging/energyLogger.py
@@ -6,8 +6,8 @@ import numpy as np
 
 from pytissueoptics.rayscattering import utils
 from pytissueoptics.rayscattering.tissues import RayScatteringScene
-from pytissueoptics.rayscattering.views.view2D import ViewGroup, View2D
-from pytissueoptics.rayscattering.views.viewFactory import ViewFactory
+from pytissueoptics.rayscattering.display.views.view2D import ViewGroup, View2D
+from pytissueoptics.rayscattering.display.views.viewFactory import ViewFactory
 from pytissueoptics.scene.logger.logger import Logger, InteractionKey
 from pytissueoptics.scene.geometry.vector import Vector
