--- pytissueoptics/rayscattering/views/__init__.py
@@ -1,3 +1,0 @@
-from .view2D import ViewGroup, View2D
-from .direction import Direction
-from .viewFactory import ViewFactory
